+++ src/Service/IotData/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## NOT RELEASED
 
+### Added
+
+- AWS api-change: changed `cn-north-1` région endpoint
+
 ## 0.1.0
 
 First version